# specialize hexadecimal sequence decoding